Allow open_timeout to be configured by ENV var
We want this value to be configurable so that it can vary by
environment. For example, GHES customers might run Dependabot in a
higher latency network. In these cases, the previous timeout of two
seconds is not enough.

Note: I had wanted to get some tests on these, but relying on constants
to store the values makes the scope of this change larger than I would
like. I'm open to re-architecting if needed. Verification of this change
would also be somewhat sticky as we'd need to reach into the Octokit
client and inspect its values. While I've done that in a debugger, doing
so in a test seems less than ideal.