fix required checks not run during a release
We have "required checks" configured for the CI, and so if the CI never runs then the build will never pass. So this creates a workflow which runs during a release with identical names to the real CI workflow, but doesn't run the real tests.

We also thought it would be a good idea to skip the CodeQL step during a release since it probably won't detect any issues. The CodeQL isn't currently required so no codeql-analysis-release.yml required yet!