Downgrade bundler to 2.3.14
The latest versions of bundler has an issue where incomplete lockfiles that are
locked to "ruby" but not including exclusively gems NOT platform specific.

This causes an issue with existing lockfiles, generating a missing gem
error.

This is explained in more detail in: https://github.com/rubygems/rubygems/pull/5807

Since this is currently preventing Dependabot updates for our users,
especially around sorbet dependencies, and since we have no good way to
signal these failures to them with an explanation on how to resolve it,
I think it's best that we downgrade bundler to a version that does not
include this, until https://github.com/rubygems/rubygems/pull/5807 is
merged and we can bump to the latest version again.