Split python tests across 5 workflows
The python tests are too damn slow. This is something that we _should_
fix at the core, by making them faster, but until we do that, let's not
waste precious time waiting on those tests to pass, let's throw more
computers at the problem.

This is a relatively crude approach, but it should result in stable
reruns. There's probably a way to make this cleaner by moving it into a
rake task or something, so we can reuse it for more ecosystems. I'll
leave that as an exercise for someone else that feels strongly about
this.