Yarn Berry: Run commands in `update-lockfile` mode
Yarn Berry commands related to performing updates can run in a mode
called `update-lockfile`, which was purpose-built for exactly what
Dependabot does:

```
- `update-lockfile` will skip the link step altogether, and only fetch packages
  that are missing from the lockfile (or that have no associated checksums).
  This mode is typically used by tools like Renovate or Dependabot to keep a
  lockfile up-to-date without incurring the full install cost.
```

This makes yarn berry updates _significantly_ more performant, but it
has the downside that when the yarn cache is committed to the repo, old
entries in there are not cleaned up. Everything still works as expected
when this happens, but over time it could lead to that cache folder
getting larger than is necessary.

I've opened an issue about this upstream:
https://github.com/yarnpkg/berry/issues/4886, not entirely sure if this
is a bug or just a consequence of how this works.

For now I think it's worth the trade-off, as many projects don't commit
the cache folder to repo, and the performance benefits are very
significant.

I'll try to find some time to address this upstream.