Yarn Berry: Prevent sub-package dependencies being added to root workspace
Previously, when attempting to update a sub-packages' dependency, it
would unintentionally get added to the top-level workspaces'
`dependencies` in the yarn lockfile.

This happened because we ran `yarn add` instead of `up`, which adds it
to the top-level package (and then updates the other existing
instances), we then discard the package.json changes, so those never got
committed, but the entry in the lockfile remains, putting it in an
inconsistent state.