Yarn Berry: Ensure multiple requirements are parsed correctly
When the lockfile contains multiple entries for a dependency that are
resolved to a single version, the lockfile will contain an entry like
so:

```
postcss@npm:^8.4.14, postcss@npm:^8.4.17, postcss@npm:^8.4.4":
  version: 8.4.17
```

This resulted in the requirements not matching, which would result in us
picking the wrong version for that dependency in some cases.

By ensuring we match these correctly, that issue is resolved.