Downgrade RubyGems to 3.3.7 to match Ruby 3.1.2
We use `dependabot-core` as a library in a few internal projects. After
the bump to `3.3.22`, they started complaining with the following error:
```
dependabot-common-0.213.0 requires rubygems version >= 3.3.22, which is
incompatible with the current version, 3.3.7
```

RubyGems `3.3.7` is the version shipped with `ruby:3.1.2`.

While we could patch all those places to update their system RubyGems,
it's a bit of a game of whack-a-mole, and adds minimal value.

So instead, relax the constraint slightly so that those versions work.

Downgrading the version used in the Dockerfile was optional, but I
thought better to keep that in sync with what most consumers of the
library will be using.