Add Security Update Support for Pub
This implements basic security update support for Pub.

For now, it's a naive implementation that bumps to the latest available
version, and then filters it out if it's vulnerable.

In other ecosystems, we only bump to the minimum version required to get
a non-vulnerable version.

We're currently discussing with the `Pub` team how best to do that over
in `https://github.com/dependabot/dependabot-core/issues/5391`, but that
will come later.