Remove unnecessary and incorrect logging
This is not actually installing any gems, just configuring Bundler to
look for gems at a specific path.

Given the new docker out (where you don't see progress by default), and
the fact that the old output shows each command being run, I don't think
we need the explicit logging so I went with removing it instead of
correcting it.