Make sure to set Bundler path at the right place
Previously, the development container would have `BUNDLE_PATH` set, so
when building it, gems would be properly installed to `omnibus/.bundle`
and all the different ecosystems would pick up gems from there.

However, since the recent split of ecosystem images, this is no longer
the case. When rebuilding and booting the development container and
trying to run tests for an ecosystem, you get the following:

```
$ bundle exec rspec

Could not find gem 'debug (~> 1.7.1)' in locally installed gems.
Run `bundle install` to install missing gems.
```

This is because `BUNDLE_PATH` is no longer set in the development
container. The reason is that the base image was split into stages and
the `BUNDLE_PATH` was left around in the "builder" stage container, not
in the final container.

The fact that nothing other than the development container was affected
shows that this ENV was only useful in the development container. So the
solution would be to move it there. However, while moving it there, I
figured it was simpler to use `bundle config path` instead of
`BUNDLE_PATH` to teach omnibus where to install & look for gems, just
like we do for the other ecosystems.

So that's what I did.